burns output in group as option
New general settings option -> Should we show burns as Group Chat (On/Off)

Will show burns in group chat instead of mq window if set to On. Sometimes while burning you want to scroll up and see what actually fired, and having the burn output in the mq2 window was a hassle since it snapped to bottom when trying to scroll.